refactor(voicecli): promote engine.py into engines/engine.py (T3 of #83)
Engine abstract class + registry promoted into the engines/ layer package.
engines/__init__.py re-exports preserve public symbols. No behavior change.
