+++ src/voicecli/daemon_protocol.py
@@ -12,7 +12,7 @@ import math
 import socket
 from pathlib import Path
 
-from voicecli.engine import QWEN_ENGINES
+from voicecli.engines.engine import QWEN_ENGINES
 
 
 _STR_MAX = 256
@@ -173,7 +173,7 @@ def sanitize_request(req: dict) -> str | None:
 
 
 def _load_engine(name: str, fast: bool = False):
-    from voicecli.engine import get_engine
+    from voicecli.engines.engine import get_engine
 
     eng = get_engine(name)
     if fast and name in QWEN_ENGINES:
@@ -183,7 +183,7 @@ def _load_engine(name: str, fast: bool = False):
 
 def _has_vram(eng_name: str) -> bool:
     """Return True if enough free VRAM is available to load the engine."""
-    from voicecli.engine import check_vram
+    from voicecli.engines.engine import check_vram
 
     try:
         check_vram(eng_name)
